Add test that hits the last page. This was useful to verify the performance improvement of https://github.com/cloudfoundry/cloud_controller_ng/pull/3983. Additionally remove the redundant `fmt.Sprintf` calls.

Since this change is already in capi-release/cf-d we won't see the delta, but running locally on mysql:

Before:
```
• [SLOW TEST] [10.332 seconds]
audit_events GET /v3/audit_events as admin getting the last page
/home/pivotal/workspace/cf-performance-tests/audit_events/v1/audit_events_test.go:128

  Begin Report Entries >>
    GET /v3/audit_events::as admin getting the last page - /home/pivotal/workspace/cf-performance-tests/audit_events/v1/audit_events_test.go:141 @ 10/24/24 16:34:00.627
      GET /v3/audit_events::as admin getting the last page
      Name                            | N | Min     | Median  | Mean    | StdDev | Max    
      ====================================================================================
      GET /v3/audit_events [duration] | 5 | 1.8773s | 1.8903s | 1.8938s | 14.5ms | 1.9114s
  << End Report Entries
------------------------------
```

After:
```
 [4.189 seconds]
audit_events GET /v3/audit_events as admin getting the last page
/home/pivotal/workspace/cf-performance-tests/audit_events/v1/audit_events_test.go:128

  Begin Report Entries >>
    GET /v3/audit_events::as admin getting the last page - /home/pivotal/workspace/cf-performance-tests/audit_events/v1/audit_events_test.go:139 @ 10/24/24 16:37:27.348
      GET /v3/audit_events::as admin getting the last page
      Name                            | N | Min     | Median  | Mean  | StdDev | Max    
      ==================================================================================
      GET /v3/audit_events [duration] | 5 | 649.4ms | 660.7ms | 661ms | 9.7ms  | 672.5ms
```